Move events handling into Redux
